commons: share MergeOptions trait
This moves `MergeOptions` from graph-builder into commons library.
